fix(stella-cli): the wrapper driver hands a plugin a real turn
The follow-through on the driver slice: `persistence.rs` had crossed the
1500-line guard carrying `TurnDoor`'s new observer, so `TurnDoor` moves to its
own module beside it. The cut follows the concern rather than the count —
everything else in `persistence` *writes* (execution rows, events, the run
terminator) while `TurnDoor` is a plain value threaded **into** a turn, holding
no store and touching no telemetry. `persistence::TurnDoor` still resolves, so
no call site moves.

`run_raw_one_shot` takes the caller's already-resolved `PipelineChoice` instead
of the variant id read out of it. That is what lets `--test-command` reach the
raw path — the argument a wrapper plugin's oracle needs to observe a flip at
all — without `agent.rs`, a grandfathered god file at its ceiling, gaining a
line: the callee does the `.plugin()` read the call site was doing.

Closes #3553
Closes #3552
Closes #3561
